move SeaResultSetQueueFactory and JsonQueue into separate SEA module
